refactor(edgeless): support configuring overlays and elementRenders
